Move UtilityAssociationGroupResultView into own file
